Destroy pixi objects to avoid graphics memory loss Minor refactoring to avoid circular dependecy
