+++ src/WellboreModule.ts
@@ -85,6 +85,11 @@ export default class WellboreModule extends ModuleInterface {
     WellboreShader.build(config.wellboreResize.max.scale, extra.wellboreDash);
   }
 
+  destroy(): void {
+    this.asyncLoop.StopAll();
+    super.destroy();
+  }
+
   registerGroup(key: string, options?: GroupOptions) : void {
     if (this.groups[key]) throw Error(`Group [${key}] already registered!`);
     this.groups[key] = new Group(key, options);
@@ -132,13 +137,13 @@ export default class WellboreModule extends ModuleInterface {
     this.containers.labels.addChild(wellbore.label.text);
     this.containers.labels.addChild(wellbore.label.background);
     group.append(wellbore);
-    root.recalculate(true);
+    root?.recalculate(true);
 
     // Add to line dictionary
     if(!wellbore.interpolator.singlePoint) this.lineDict.add(projectedPath, wellbore);
 
     // Append wellbore to root
-    root.append(wellbore);
+    root?.append(wellbore);
 
     if (this._deferredSelector && this._deferredSelector(wellbore.data)) {
       this._deferredSelector = undefined;
@@ -351,9 +356,9 @@ export default class WellboreModule extends ModuleInterface {
     this.roots = [];
 
     // remove PIXI elements
-    this.containers.wellbores.removeChildren();
-    this.containers.labels.removeChildren();
-    this.containers.roots.removeChildren();
+    this.containers.wellbores.removeChildren().forEach(child => child.destroy());
+    this.containers.labels.removeChildren().forEach(child => child.destroy());
+    this.containers.roots.removeChildren().forEach(child => child.destroy());
     this.pixiOverlay.redraw();
   }
 
@@ -381,9 +386,9 @@ export default class WellboreModule extends ModuleInterface {
         }
 
         // remove PIXI elements
-        this.containers.wellbores.removeChild(w.mesh);
-        this.containers.labels.removeChild(w.label.text);
-        this.containers.labels.removeChild(w.label.background);
+        w.mesh?.destroy();
+        w.label?.text?.destroy();
+        w.label?.background?.destroy();
       });
       group.wellbores = [];
     });
